Wip: update message actions
