+++ app/message-pin/client/actionButton.js
@@ -13,7 +13,7 @@ Meteor.startup(function() {
 	MessageAction.addButton({
 		id: 'pin-message',
 		icon: 'pin',
-		label: 'Pin_Message',
+		label: 'Pin',
 		context: ['pinned', 'message', 'message-mobile'],
 		action() {
 			const { msg: message } = messageArgs(this);
@@ -31,14 +31,14 @@ Meteor.startup(function() {
 
 			return hasAtLeastOnePermission('pin-message', message.rid);
 		},
-		order: 20,
+		order: 7,
 		group: 'menu',
 	});
 
 	MessageAction.addButton({
 		id: 'unpin-message',
 		icon: 'pin',
-		label: 'Unpin_Message',
+		label: 'Unpin',
 		context: ['pinned', 'message', 'message-mobile'],
 		action() {
 			const { msg: message } = messageArgs(this);
@@ -56,7 +56,7 @@ Meteor.startup(function() {
 
 			return hasAtLeastOnePermission('pin-message', message.rid);
 		},
-		order: 21,
+		order: 8,
 		group: 'menu',
 	});
 
@@ -85,7 +85,7 @@ Meteor.startup(function() {
 	MessageAction.addButton({
 		id: 'permalink-pinned',
 		icon: 'permalink',
-		label: 'Permalink',
+		label: 'Get_link',
 		classes: 'clipboard',
 		context: ['pinned'],
 		async action(event) {
