+++ app/threads/client/index.js
@@ -4,3 +4,5 @@ import './flextab/threads';
 import './threads.css';
 import './messageAction/follow';
 import './messageAction/unfollow';
+import './messageAction/replyInThread';
+
