+++ app/ui-utils/client/lib/popover.js
@@ -3,13 +3,10 @@ import { Meteor } from 'meteor/meteor';
 import { Blaze } from 'meteor/blaze';
 import { FlowRouter } from 'meteor/kadira:flow-router';
 import { Template } from 'meteor/templating';
-import { TAPi18n } from 'meteor/tap:i18n';
-import { isRtl, handleError } from '../../../utils';
-import { messageArgs } from './messageArgs';
-import { ChatSubscription } from '../../../models';
+import { isRtl, handleError } from '../../../utils/client';
+import { ChatSubscription } from '../../../models/client';
 import _ from 'underscore';
 import { hide, leave } from './ChannelActions';
-import { modal } from './modal';
 import { messageBox } from './messageBox';
 import { MessageAction } from './MessageAction';
 import { RoomManager } from './RoomManager';
@@ -177,42 +174,6 @@ Template.popover.events({
 			button.action.call(t.data.data, e, t.data.instance);
 			popover.close();
 			return false;
-		}
-
-		if (e.currentTarget.dataset.id === 'report-abuse') {
-			const { msg: message } = messageArgs(t.data.data);
-			modal.open({
-				title: TAPi18n.__('Report_this_message_question_mark'),
-				text: message.msg,
-				inputPlaceholder: TAPi18n.__('Why_do_you_want_to_report_question_mark'),
-				type: 'input',
-				showCancelButton: true,
-				confirmButtonColor: '#DD6B55',
-				confirmButtonText: TAPi18n.__('Report_exclamation_mark'),
-				cancelButtonText: TAPi18n.__('Cancel'),
-				closeOnConfirm: false,
-				html: false,
-			}, (inputValue) => {
-				if (inputValue === false) {
-					return false;
-				}
-
-				if (inputValue === '') {
-					modal.showInputError(TAPi18n.__('You_need_to_write_something'));
-					return false;
-				}
-
-				Meteor.call('reportMessage', message._id, inputValue);
-
-				modal.open({
-					title: TAPi18n.__('Report_sent'),
-					text: TAPi18n.__('Thank_you_exclamation_mark '),
-					type: 'success',
-					timer: 1000,
-					showConfirmButton: false,
-				});
-			});
-			popover.close();
 		}
 	},
 	'click [data-type="sidebar-item"]'(e, instance) {
