+++ imports/message-read-receipt/client/room.js
@@ -6,12 +6,12 @@ import { settings } from '../../../app/settings';
 MessageAction.addButton({
 	id: 'receipt-detail',
 	icon: 'info-circled',
-	label: 'Message_info',
+	label: 'Info',
 	context: ['starred', 'message', 'message-mobile'],
 	action() {
 		const { msg: message } = messageArgs(this);
 		modal.open({
-			title: t('Message_info'),
+			title: t('Info'),
 			content: 'readReceipts',
 			data: {
 				messageId: message._id,
@@ -24,6 +24,6 @@ MessageAction.addButton({
 	condition() {
 		return settings.get('Message_Read_Receipt_Store_Users');
 	},
-	order: 1,
+	order: 10,
 	group: 'menu',
 });
