+++ packages/rocketchat-i18n/i18n/en.i18n.json
@@ -655,6 +655,7 @@
   "Convert_Ascii_Emojis": "Convert ASCII to Emoji",
   "Copied": "Copied",
   "Copy": "Copy",
+  "Copy_text": "Copy Text",
   "Copy_to_clipboard": "Copy to clipboard",
   "COPY_TO_CLIPBOARD": "COPY TO CLIPBOARD",
   "could-not-access-webdav": "Could not access WebDAV",
@@ -1416,6 +1417,7 @@
   "Full_Screen": "Full Screen",
   "Gaming": "Gaming",
   "General": "General",
+  "Get_link": "Get Link",
   "github_no_public_email": "You don't have any email as public email in your GitHub account",
   "Give_a_unique_name_for_the_custom_oauth": "Give a unique name for the custom oauth",
   "Give_the_application_a_name_This_will_be_seen_by_your_users": "Give the application a name. This will be seen by your users.",
@@ -1926,6 +1928,8 @@
   "Mark_all_as_read": "Mark all messages (in all channels) as read",
   "Mark_as_read": "Mark As Read",
   "Mark_as_unread": "Mark As Unread",
+  "Mark_read": "Mark Read",
+  "Mark_unread": "Mark Unread",
   "Markdown_Headers": "Allow Markdown headers in messages",
   "Markdown_Marked_Breaks": "Enable Marked Breaks",
   "Markdown_Marked_GFM": "Enable Marked GFM",
@@ -2230,6 +2234,7 @@
   "Permissions": "Permissions",
   "Personal_Access_Tokens": "Personal Access Tokens",
   "Phone_number": "Phone number",
+  "Pin": "Pin",
   "pin-message": "Pin Message",
   "pin-message_description": "Permission to pin a message in a channel",
   "Pin_Message": "Pin Message",
@@ -2406,7 +2411,10 @@
   "Replied_on": "Replied on",
   "Replies": "Replies",
   "Reply": "Reply",
+  "Reply_directly": "Reply Directly",
+  "Reply_in_thread": "Reply in Thread",
   "ReplyTo": "Reply-To",
+  "Report": "Report",
   "Report_Abuse": "Report Abuse",
   "Report_exclamation_mark": "Report!",
   "Report_sent": "Report sent",
@@ -2685,6 +2693,7 @@
   "Sound": "Sound",
   "Sound_File_mp3": "Sound File (mp3)",
   "SSL": "SSL",
+  "Star": "Star",
   "Star_Message": "Star Message",
   "Starred_Messages": "Starred Messages",
   "Start": "Start",
@@ -2930,6 +2939,7 @@
   "Unmute_someone_in_room": "Unmute someone in the room",
   "Unmute_user": "Unmute user",
   "Unnamed": "Unnamed",
+  "Unpin": "Unpin",
   "Unpin_Message": "Unpin Message",
   "Unread": "Unread",
   "Unread_Count": "Unread Count",
